# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateUserAPIKeyRequest:
    properties:
      api_key:
        description: The actual API key
        type: string
      name:
        description: User-friendly name for the key
        type: string
      provider:
        description: Provider name (gemini, openai, anthropic)
        type: string
    required:
    - api_key
    - name
    - provider
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        description: |-
          HTTP status code
          example: 400
        example: 400
        type: integer
      details:
        description: Field-level validation error details (for validation errors)
        items:
          $ref: '#/definitions/models.FieldError'
        type: array
      error:
        description: |-
          Error type/category
          example: Bad Request
        example: Bad Request
        type: string
      message:
        description: |-
          Detailed error message
          example: Invalid email format
        example: Invalid email format
        type: string
      request_id:
        description: |-
          Request ID for tracing and debugging
          example: 550e8400-e29b-41d4-a716-446655440000
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  models.FieldError:
    properties:
      code:
        description: |-
          Error code for programmatic handling
          example: email
        type: string
      field:
        description: |-
          The field name that failed validation
          example: email
        type: string
      message:
        description: |-
          Human-readable error message
          example: Invalid email format
        type: string
      value:
        description: The rejected value (excluded for sensitive fields)
    type: object
...
  models.SuccessResponse:
    properties:
      data:
        description: Response data (varies by endpoint)
      message:
        description: |-
          Success message
          example: User created successfully
        example: User created successfully
        type: string
      success:
        description: |-
          Operation success status
          example: true
        example: true
        type: boolean
    type: object
...
  models.UpdateUserAPIKeyRequest:
    properties:
      api_key:
        description: New API key value
        type: string
      is_active:
        description: Whether the key is active
        type: boolean
      name:
        description: New name for the key
        type: string
    type: object
...
  models.UserAPIKeyResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      key_preview:
        type: string
      last_used_at:
        type: string
      name:
        type: string
      provider:
        type: string
      updated_at:
        type: string
      usage_count:
        type: integer
    type: object
  models.UserAPIKeysResponse:
    properties:
      count:
        type: integer
      keys:
        items:
          $ref: '#/definitions/models.UserAPIKeyResponse'
        type: array
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/your-org/react-golang-starter-kit/issues
  description: A comprehensive REST API for the React Go Starter Kit application built
    with Fiber, GORM, and PostgreSQL. This API provides secure user authentication,
    user management, and system health monitoring.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/your-org/react-golang-starter-kit
  title: React Go Starter Kit API
  version: 1.0.0
paths:
...
  /api/users/me/api-keys:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserAPIKeysResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user API keys
      tags:
      - User Settings
    post:
      parameters:
      - description: API key details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserAPIKeyRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserAPIKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Key already exists for this provider
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create user API key
      tags:
      - User Settings
  /api/users/me/api-keys/{id}:
    delete:
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user API key
      tags:
      - User Settings
    get:
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserAPIKeyResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user API key by ID
      tags:
      - User Settings
    put:
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      - description: API key updates
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserAPIKeyRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserAPIKeyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user API key
      tags:
      - User Settings
  /api/users/me/api-keys/{id}/test:
    post:
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Test user API key
      tags:
      - User Settings
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Format: `Authorization:
      Bearer {token}`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"